Group gomod digest updates
The golang.org modules (golang.org/x/foo) are essentially always
updated. With the current configuration of Mon/Fri updates an no
grouping it's like a game of whack-a-mole to get `seek-auth-sidecar`'s
PRs merged.

This treats digest updates like `aws-sdk` and only updates digests on
the first of the month with grouping.