Bump `prConcurrentLimit` to 9 temporarily
We're seeing issues with the concurrency limit being enforced against
the branch count rather than the PR count. This is a hammer to push PRs
through while the Renovate team are investigating.

This may lead to a bit of PR spam but I'm hoping that this will mostly
surface branches that were already created but not yet PRed.