Spell out `h:mm` in schedules
Landing #25 required a small battle with `@breejs/later` to get it to
recognise a fortnightly schedule. It seems the minutes are the key to
getting the text parser to work here.

This opens up the door for us to bump a few other things to a
fortnightly schedule, but I'll leave that to another PR.