Try to force `prCreation` for trusted deps
We seem to be receiving far fewer Renovate PRs as a result of #54. This
includes SEEK deps where I've had to manually go to Dependency
Dashboards to force their creation.

My working theory is that a deferred `prCreation` means that Renovate
has to take additional passes at a given branch to get it PRed, and that
may be leading to some internal rate limiting. Immediate PR creation
means that the work can be bundled alongside the initial commit and
branch creation.