Revert range strategy to "replace" (#83)
* Revert range strategy to "replace"

Per renovatebot/renovate#19800, the "auto" strategy now raises
additional PRs to update dependencies in range. This is either
unsuitable for our use case or bugged, as we have "lockfile update" PRs
(as denoted by a `-lockfile` branch suffix) which are bumping the
in-range version of a package's production dependencies. This means that
e.g. `^2.0.0` gets bumped to `^2.1.0` and so on, and results in a lot of
transitive noise for projects that consume the package.

This _may_ break the previous behaviour of widening peer dependency
ranges, but's a lesser concern now and I'm following up on the Renovate
recommendation for handling this use case.

https://docs.renovatebot.com/release-notes-for-major-versions/#version-35

https://github.com/renovatebot/renovate/discussions/20990

* Widen peer deps