Restore gitIgnoredAuthors (#126)
I reverted this in https://github.com/seek-oss/rynovate/pull/125 because I was noticing an autofix-rebase loop with a skuba upgrade.

I have traced this down to skuba's upgrade changing Docker images. I'm a little confused on the details, but Renovate then decided the branch was out of date because there were new dependencies to update.

I believe that restoring skuba to be in its own PR will bypass this issue.